+++ Storage.js
@@ -1,16 +1,18 @@
 
-// set value of key in local storage
-export function set_local_storage(key, value) {
-	localStorage.setItem(key, value);
-}
 
-// get value of key in local storage as json
-export function get_local_storage(key) {
-	return JSON.parse(localStorage.getItem(key));
-}
+export class LocalStorage {
+	static set(key, value) { 
+		if (typeof value === 'object') 
+			localStorage.setItem(key, JSON.stringify(value));
+		else 
+			localStorage.setItem(key, value.toString());
+	}
+	static get_json(key) { return JSON.parse(localStorage.getItem(key)); }
+	static get_string(key) { return localStorage.getItem(key); }
+	static get_number(key) { return parseFloat(localStorage.getItem(key)); }
+	static get_boolean(key) { return localStorage.getItem(key) === 'true'; }
 
-// remove key from local storage
-export function remove_local_storage(key) {
-	localStorage.removeItem(key);
-}
+	static remove(key) { localStorage.removeItem(key); }
+	static clear() { localStorage.clear(); }
 
+}